wip: adding config folder
