+++ libs/shared/config/src/index.ts
@@ -1,1 +1,1 @@
-export * from './lib/shared-config'
+export * from './env'
